requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
whitenoise==6.7.0
